:truck: move AnimatorsMapSerialize test
js doesn't detect the difference between int and double
